+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed ordering of study collections in the 'Add Study to Collection' dialog (#125).
 - Fixed refresh of study collection table after adding a study to a collection (#126).
 - Fixed obstructed CRO dropdown in the study form (#123).
+- User table correctly refreshes on status change (#140).
 
 ## [0.9.4] - 2023-10-20